Add files via upload
These HTML files are parts of a starter website that I started to develop last week and over the weekend. Please excuse me that it is in Hungarian, but I'm guessing that it's not the text that is important but the look and function of the site. Please note that I started to experiment with "Balatonalmadi.html" to make it responsive with the use of flex boxes.